+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 
+- Authn Menu - Remove documentation, support and invite user entries
+- Authn Menu - Dropdown changed to MenuItems
 - Nuvlabox detail page - hide generate-new-api-key
 - Session - Error message in sign-in and sign-up pages not always
   visible [#778](https://github.com/nuvla/ui/issues/778)